requirements.txt
streamlit==1.40.1

python-dotenv
pydantic

langchain
langchain-core

langchain-groq